Improvements to autonamy and ease of addition/modification
- Alfred is smarter than ever. Great care has been taken to remove any
direct references to specific language modules so that new modules can
be added in XML only(!) meaning that no code has to be written into the
project solution. (I hope to eventually extend this functionality to the
"functions" that are triggered by context buttons so that these can be
added on the fly).

- I'm experimenting with a new look and this starts here, the ability to
play animated GIF's has been built in and is used to make the backdrop
more visually appealing, (perhaps the days are numbered for the original
mustache look).

- Massive improvements to the back end of the UI, replies can now be
scrolled (!) and the interface is happy to deal with long replies.

- Contextual buttons are now smarter, you are now able to use one
button, two, all or none without any issues. To omit a button, a "-"
should be put inside the XML tags.

- A 'debugging' mode has been added, essentially this is just an
override to the voice input and allows me to work on the system away
from my PC.  I'm sure I'll improve this as I get frustrated with having
to rebuild the project to change my input...

- A few new error handling features, crashes should be very rare now.