Weather API - darksky.net
- Weather API has been configured and is operational on the mobile app.
It's current implementation retrieves 17 pieces of data regarding the
users current location (determined by longitude and latitude
coordinates).

- Tasks still to complete include the transferring of this data to the
watch and presenting it to the user.

- Watch -> Phone communication is still not happening, this is a big
problem and needs urgent attention.  This has not stopped me working on
other areas of the app as I am able to create manual workarounds
(buttons that I press in the main app when the watch is waiting for a
response)