Data Retrieval Automation
- The API data retrieval process is almost fully automated. The watch
sends a string to the phone which will correspond to an API to retrieve
data from. The phone then retrieves that data and sends it back to the
watch where it is parsed and saved in Shared Preferences.

- At this point there is no communciation between the listener on the
watch and its main class to inform it that the data is ready to be read
and the main class will simply read shared preferences before it is
ready, essentially always retrieving the last data.