Google Calendar API Integration
- Calendar is now integrated (although a little rough around the edges).
It works along the same lines as a weather api call however the phone
will direct it into a seperate class. At the moment for this to work the
user has to manually grant all permissions in their settings menu
(Android 6.0+ Permissions are a pain to work with...) and then select
the Google account that they want to use with Alfred on the phone (no
way around this by the looks of it).

- Throughout implementing Calendar a few optimizations and processes
have been simplified and smoothed out.

- I think that the data layer transmission on the phone needs to be in
its own class in order to avoid repetition and to allow good OO design
(this may be a large task so I will leave it until the end)