Beginnings of Input Modification Intent
- This intent will be called after the voice recognition tool returning
the voice input data but before this input has been optimised and
analysed. The idea being that the intent will appear and allow the user
three seconds to make any changes to their input before it is confirmed
by the system. Initially this will just be the ability to cancel or
re-record the input although I hope to expand this to changing
individual words in the input without having to re-record the entire
thing. (No features currently functional)