Repaired the broken Voice Recognition service
- During the addition of a previous feature that required an
ActivityResult callback, I had to move and modify the voice dictation
callback. This led to the system becoming broken and duplicating
requests for special functions, leading to the system misreading the
input it was recieving and asking the phone for data that did not match
anything it knew about. This has been fixed and voice dictation should
work perfectly again.