Wikipedia API
- Wikipedia API has been added for reference questions along the lines
of "Who is...", "What is..." "Who were..." etc. The API call retrieves
the page ID (could do something with this later), the title and the page
introduction. The introduction is the only piece of information that is
currently used as it is displayed back to the user as a response.